Drop support of Travis and AppVeyor.
Move to Github Actions instead.

This also updates everything to the latest best practices, and
simplifies the template as much as possible:

* Don't let black, isort, pre-commit be optional
* Only use codecov for coverage
* Don't support RTD
* Don't support Jupyter notebooks in docs
* Don't support git-flow

All these options were making the template extremely difficult to
maintain. Support for Jupyter notebooks is still too much in flux at
this point, and also a pretty special use case.